requirements.txt
# Dependências do projeto

psycopg2-binary
